backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
motor>=3.3.2
pymongo>=4.6.3
pydantic>=2.0.0
python-dotenv>=1.0.1
google-genai>=1.0.0
httpx>=0.27.0
